SWAP-1075 Rename ProposalComponent...(s) to QuestionaryComponent...(s)
