+++ src/components/questionary/formComponents/QuestionaryComponentProposalBasis.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+import { BasicComponentProps } from 'components/proposal/IBasicComponentProps';
+
+function QuestionaryComponentProposalBasis(props: BasicComponentProps) {
+  return <div></div>;
+}
+
+export default QuestionaryComponentProposalBasis;
